Move test-only ChainSyncStatus code into its own module
